split routes into router
