+++ routes/server.js
@@ -4,16 +4,21 @@ if (require.main === module) {
 
 const express = require('express')
 const mongoose = require('mongoose')
-const uuidv4 = require('uuid/v4')
 const bodyparser = require('body-parser')
 const passport = require('passport')
 const jwt = require('jsonwebtoken')
 
-const { Course } = require('../models/course')
-const { User } = require('../models/user')
 const { jwtStrategy } = require('../strategies')
 const { DATABASE_URL, PORT } = require('../config.js')
 const app = express()
+
+const coursesRouter = require('./coursesRouter')
+const draftsRouter = require('./draftsRouter')
+const usersRouter = require('./usersRouter')
+
+app.use('/api/courses', coursesRouter)
+app.use('/api/drafts', draftsRouter)
+app.use('/api/users', usersRouter)
 
 mongoose.Promise = global.Promise
 
@@ -22,10 +27,6 @@ app.use('/node_modules', express.static('node_modules'))
 app.use(bodyparser.json())
 
 passport.use(jwtStrategy)
-
-const jwtAuth = passport.authenticate('jwt', { session: false })
-
-// ~~~~~~ http requests ~~~~~~
 
 app.get('/course/:courseId', (req, res) => {
   const options = {
@@ -42,116 +43,6 @@ app.get('/create/:courseId', (req, res) => {
 
   res.sendFile('create.html', options)
 })
-
-// ~~~~~~ DRAFTS API ~~~~~~
-
-// create function to make new draft and update user
-async function createNewDraftAndUpdateUser(draft, userId) {
-  const newDraft = { ...draft, courseId: uuidv4() }
-
-  const updatedUser = await User.findOneAndUpdate(
-    { userId },
-    { $push: { drafts: newDraft } },
-    { new: true }
-  )
-
-  return updatedUser
-}
-
-app.post('/api/drafts', jwtAuth, async (req, res) => {
-  try {
-    // don't have userId in URL, get it from JWT
-
-    // need to send userId with update
-    // make sure user has access to do this
-    // req.user should have jwt info
-    const newDraft = await createNewDraftAndUpdateUser(req.body, req.user._id)
-    res.send(newDraft)
-  } catch (err) {
-    console.error(err)
-  }
-})
-
-async function updateDraftInUserObject(updatedDraft, userId) {
-  const userToUpdate = await User.findOne({ userId: userId })
-
-  const draftToUpdate = userToUpdate.drafts.find(
-    draft => draft.courseId === updatedDraft.courseId
-  )
-
-  if (draftToUpdate) {
-    draftToUpdate.remove()
-    userToUpdate.drafts.push(updatedDraft)
-  }
-
-  await userToUpdate.save()
-
-  return updatedDraft
-}
-
-app.put('/api/drafts', jwtAuth, async (req, res) => {
-  try {
-    await updateDraftInUserObject(req.body, req.user._id)
-    res.send(req.body)
-  } catch (err) {
-    console.error(err)
-  }
-})
-
-// ~~~~~~ USER API ~~~~~~
-
-async function createNewUser(userData) {
-  const newUser = await User.create(userData)
-
-  const token = jwt.sign({ user: { _id: newUser._id } }, process.env.JWT_SECRET)
-  const user = {
-    jwt: token,
-    userData: newUser
-  }
-
-  return user
-}
-
-app.post('/api/users', jwtAuth, async (req, res) => {
-  try {
-    const newUser = await createNewUser(req.body)
-    res.send(newUser)
-  } catch (err) {
-    console.error(err)
-  }
-})
-
-async function getUser(userId) {
-  return await User.findOne({ userId: userId })
-}
-
-app.get('/api/users/:userId', async (req, res) => {
-  // can i just use mongo id instead of creating one with uuid?
-  try {
-    const user = await getUser(req.params.userId)
-    res.send(user)
-  } catch (err) {
-    console.error(err)
-  }
-})
-
-// ~~~~~~ COURSES API ~~~~~~
-
-async function publishCourse(course) {
-  await Course.findOne({ courseId: course.courseId }).remove()
-  return await Course.create(course)
-}
-
-app.post('/api/courses', jwtAuth, async (req, res) => {
-  try {
-    const newCourse = await publishCourse(req.body)
-    res.send(newCourse)
-  } catch (err) {
-    console.error(err)
-  }
-})
-
-// ~~~~~~ start and stop server ~~~~~~
 
 let server
 
@@ -195,10 +86,5 @@ if (require.main === module) {
 module.exports = {
   runServer,
   app,
-  closeServer,
-  createNewDraftAndUpdateUser,
-  updateDraftInUserObject,
-  createNewUser,
-  getUser,
-  publishCourse
+  closeServer
 }
